Fixed ESC

+++ utils/conversion.py
@@ -15,4 +15,4 @@ class Converter:
         self.__SCALE = (self.__TARGET_MAX - self.__TARGET_MIN) / (self.__SOURCE_MAX - self.__SOURCE_MIN)
 
     def getTargetValue(self, number) -> int:
-        return (self.__TARGET_MAX - ((self.__SOURCE_MAX - number) * self.__SCALE))
+        return int((self.__TARGET_MAX - ((self.__SOURCE_MAX - number) * self.__SCALE)))
